requirements.txt
flask
tensorflow-cpu ; python_version < '3.12'
pillow
